requirements.txt
langchain>=0.1.0
langchain-openai>=0.3.12
langgraph>=0.3.24
openai>=1.68.2
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
tiktoken>=0.7.0 